pyproject.toml
[project]
name = "rag-document-search"
version = "0.1.0"
description = "Beginner RAG project with PyTorch CUDA 13.0 and cuDNN 9.1"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "beautifulsoup4",
    "sentence-transformers>=5.2.0",
    "torch",
    "torchaudio",
    "torchvision",
    "huggingface-hub>=0.36.0",
    "ipywidgets>=8.1.8",
    "jq>=1.10.0",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pip>=25.3",
    "pydantic",
    "python-dotenv>=1.2.1",
    "python-docs>=0.1.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "langchain-experimental>=0.4.1",
    "faiss-cpu>=1.13.2",
    "langgraph>=1.0.6",
    "langgraph-cli[inmem]>=0.4.12",
    "wikipedia",
    "chainlit",
    "requests>=2.32.5",
]

# ─────────────────────────────
# CUDA 13 PyTorch Index (uv)
# ─────────────────────────────
[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

# ─────────────────────────────
# Package → Index Mapping
# ─────────────────────────────
[tool.uv.sources]
torch = [{ index = "pytorch-cu130" }]
torchvision = [{ index = "pytorch-cu130" }]
torchaudio = [{ index = "pytorch-cu130" }]
